added db.go with functions to add item to table and list files in table
